requirements.txt
torch
torchvision
wandb
matplotlib
numpy 
Pillow
tqdm